Refactor middleware/AuthUser.js
