--- middleware/authMiddleware.js
@@ -1,30 +1,0 @@
-const jwt = require('jsonwebtoken');
-
-const authMiddleware = (req, res, next) => {
-  const token = req.headers.authorization;
-
-  if (!token) {
-    return res.status(401).json({ message: 'Unauthorized' });
-  }
-
-  try {
-    const decoded = jwt.verify(token.split(' ')[1], process.env.JWT_SECRET);
-    req.user = decoded;
-    next();
-  } catch (error) {
-    return res.status(401).json({ message: 'Invalid token' });
-  }
-};
-
-const isAdminMiddleware = (req, res, next) => {
-  if (req.user.role === 'admin') {
-    next();
-  } else {
-    return res.status(403).json({ message: 'Forbidden' });
-  }
-};
-
-module.exports = {
-  authMiddleware,
-  isAdminMiddleware
-};
