+++ routes/transactionRoutes.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const transactionController = require('../controllers/transactionController');
-const authMiddleware = require('../middleware/authMiddleware');
+const authMiddleware = require('../middleware/AuthUser');
 
 router.post('/', authMiddleware, transactionController.createTransaction);
 router.get('/', authMiddleware, transactionController.getTransactionsByUserId);
